refactor(desktop): 抽离 workspace xterm 工厂并引入终端 addon 依赖
将 Shell 标签内嵌终端逻辑迁入 workspace-xterm 模块，并添加
@xterm/addon-webgl 与 @xterm/addon-web-links 供后续 Phase 使用。
