+++ apps/desktop/package.json
@@ -43,6 +43,8 @@
     "@spirit-agent/agent-core": "file:../../packages/agent-core",
     "@spirit-agent/host-internal": "file:../../packages/host-internal",
     "@xterm/addon-fit": "^0.10.0",
+    "@xterm/addon-web-links": "^0.11.0",
+    "@xterm/addon-webgl": "^0.18.0",
     "@xterm/xterm": "^5.5.0",
     "@xyflow/react": "^12.10.2",
     "class-variance-authority": "^0.7.1",
